+++ lang/de/bookings.php
@@ -2,7 +2,7 @@
 
 return [
   'booking_create_no_mail' =>
-    'Die Buchung wurde erfolgreich generiert, die Bestätigungsemail konnte jedoch nicht verschickt werden.',
+  'Die Buchung wurde erfolgreich generiert, die Bestätigungsemail konnte jedoch nicht verschickt werden.',
   'booking_create_success' => 'Buchung erfolgreich generiert.',
   'error_creating_booking' => 'Beim Erstellen der Buchung ist ein Fehler aufgetreten.',
   'single_occupancy' => 'Diese Buchung ist eine Einzelbelegung.',
@@ -14,7 +14,7 @@ return [
   'email_already_taken' => 'Ein Passagier mit dieser eMailadresse wurde bereits dieser Buchung hinzugefügt.',
   'pax_order_taken' => 'Die Reihenfolge der Passagiere ist bereits aufgenommen.',
   'use_manual_add_instead' =>
-    'Diese eMailadresse wurde bereits in einer anderen Buchung benutzt, nutze stattdessen die manuelle Eingabe.',
+  'Diese eMailadresse wurde bereits in einer anderen Buchung benutzt, nutze stattdessen die manuelle Eingabe.',
   'email_already_invited' => 'Ein Passagier mit dieser eMailadresse wurde bereits zu dieser Buchung eingeladen.',
   'slot_taken' => 'Der Passagier-Slot ist bereits vergeben.',
   'invite_sent' => 'Einladung wurde versendet.',
@@ -28,4 +28,12 @@ return [
   'empty_seat_removed' => 'Leeres Bett entfernt',
   'already_has_booking' => 'Du hast bereits eine bestehende Buchung für dieses Event.',
   'already_has_invitation' => 'Du wurdest bereits zu einer Buchung für dieses Event eingeladen.',
+  'error' => [
+    'email_required' => 'eMailadresse ist erforderlich',
+    'cannot_invite_yourself' => 'Du kannst dir selbst keine Einladung per eMailadresse schicken. Wenn du dieselbe eMailadresse für einen anderen Passagier verwenden möchtest, fülle das Formular bitte manuell aus.',
+    'single_occupancy' => 'Du kannst keine weiteren Passagiere zu einer Einzelbelegung hinzufügen.',
+    'email_already_used' => 'Diese eMailadresse wird bereits in einer anderen Buchung für dieses Event verwendet. Wenn du dieselbe eMailadresse erneut nutzen möchtest, fülle bitte das Formular manuell aus.',
+    'passenger_order_taken' => 'Für diesen Platz wurde bereits ein Passagier hinzugefügt. Bitte lade die Seite neu oder wähle einen anderen Platz.',
+    'email_already_invited' => 'Ein Passagier mit dieser eMailadresse wurde bereits zu dieser Buchung eingeladen.',
+  ],
 ];